workers/constructors/SharedWorker/number-arguments.html

<!doctype html>
<title>1 as arguments</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(function() {
  const worker = new SharedWorker(1, 1);
  worker.port.onmessage = this.step_func_done(function(e) {
    assert_equals(e.data[0], '1', 'first arg');
    assert_equals(e.data[1], '1', 'second arg');
  });
}, 'Test constructing a shared worker with 1');
</script>
